requirements.txt
cmake==3.18.4.post1
dlib==19.21.1
numpy==1.20.1
opencv-python==4.5.1.48
pynput==1.7.2
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pywin32==300
six==1.15.0
